+++ GUI/Model/SnakeROIModel.h
@@ -78,12 +78,12 @@ public:
 
   void SetParent(GenericSliceModel *);
 
-  bool ProcessPushEvent(float x, float y);
+  bool ProcessPushEvent(double x, double y);
 
   bool ProcessDragEvent(
-      float x, float y, float xStart, float yStart, bool release);
+      double x, double y, double xStart, double yStart, bool release);
 
-  bool ProcessMoveEvent(float x, float y);
+  bool ProcessMoveEvent(double x, double y);
 
   void ProcessLeaveEvent();
   void ProcessEnterEvent();
@@ -100,7 +100,7 @@ protected:
 
   // Four vertices in the region box (correspond to the two corners
   // of the 3D region of interest
-  Vector3f m_CornerDragStart[2];
+  Vector3d m_CornerDragStart[2];
 
   /**
    * The four edges in the rectangle, ordered first by orientation
@@ -111,23 +111,23 @@ protected:
   SnakeROISideSelectionState m_Highlight;
 
   /** Map from system's ROI in image coordinates to 2D slice coords */
-  void GetSystemROICorners(Vector3f corner[2]);
+  void GetSystemROICorners(Vector3d corner[2]);
 
   /** Compute the slice-space vertices corresponding to an edge */
   void GetEdgeVertices(unsigned int direction,
-    unsigned int index,Vector2f &x0,Vector2f &x1,const Vector3f corner[2]);
+    unsigned int index,Vector2d &x0,Vector2d &x1,const Vector3d corner[2]);
 
   /** Compute a distance to an edge */
-  float GetEdgeDistance(unsigned int direction,
-    unsigned int index,const Vector2f &point,const Vector3f corner[2]);
+  double GetEdgeDistance(unsigned int direction,
+    unsigned int index,const Vector2d &point,const Vector3d corner[2]);
 
   /**
    * Update the region of interest in response to the dragging or release
    * operations.
    */
-  void UpdateCorners(const Vector2f &xPress, const Vector2f &xDrag);
+  void UpdateCorners(const Vector2d &xPress, const Vector2d &xDrag);
 
-  SnakeROISideSelectionState ComputeSelection(Vector2f &uvSlice, Vector3f corners[]);
+  SnakeROISideSelectionState ComputeSelection(Vector2d &uvSlice, Vector3d corners[]);
 
   // Parent model
   GenericSliceModel *m_Parent;
