+++ GUI/Renderer/GenericSliceRenderer.h
@@ -149,7 +149,7 @@ protected:
   // List of child renderers
   std::list<AbstractRenderer *>m_ChildRenderers;
 
-  Vector3f ComputeGridPosition(const Vector3f &disp_pix, const itk::Index<2> &slice_index, ImageWrapperBase *vecimg);
+  Vector3d ComputeGridPosition(const Vector3d &disp_pix, const itk::Index<2> &slice_index, ImageWrapperBase *vecimg);
 };
